+++ src/SsgContext.ts
@@ -1,10 +1,6 @@
 import {SsgFile} from "./util/file/SsgFile"
 import {Logger} from "./Logger"
 
-export type BuiltInVars = SsgFile | undefined
-
-export type VarProp<V> = keyof BuiltInVars | keyof V
-
 export interface SsgContext<V = any> extends Logger {
   /**
    * Context identifier. Used in the logs.
@@ -33,9 +29,9 @@ export interface SsgContext<V = any> extends Logger {
    */
   outputFile: SsgFile
 
-  getVar(varName: VarProp<V>): string | undefined
+  getVar(varName: string): string | undefined
 
-  setVar(varName: VarProp<V>, value: any): void
+  setVar(varName: string, value: any): void
 
   clone(): SsgContext<V>