+++ src/SsgContextImpl.test.ts
@@ -2,20 +2,28 @@ import {SsgContextImpl} from "./SsgContextImpl"
 
 interface MyVars {
   someVar?: string
+  name?: string
 }
 
 describe("SsgContextImpl", () => {
 
   test("sets and gets variable", () => {
-    const myVars: MyVars = {someVar: undefined}
+    const myVars = new Map([["someVar", undefined]])
     const context = new SsgContextImpl<MyVars>("fr", myVars)
     const value = "someValue"
     context.setVar("someVar", value)
     expect(context.getVar("someVar")).toBe(value)
   })
 
+  test("sets and gets context variable", () => {
+    const myVars = new Map([["_name", "some name"]])
+    const context = new SsgContextImpl("fr", myVars)
+    expect(context.getVar("_name")).toBeDefined()
+    expect(context.getVar("$context._name")).toBe("Ssg")
+  })
+
   test("change name", () => {
-    const context = new SsgContextImpl<MyVars>("fr", {})
+    const context = new SsgContextImpl<MyVars>("fr")
     expect(context.name).toBe(SsgContextImpl.DEFAULT_NAME)
     let newName = "New name"
     context.name = newName
@@ -24,7 +32,7 @@ describe("SsgContextImpl", () => {
   })
 
   test("push and pop", () => {
-    const context = new SsgContextImpl<MyVars>("fr", {})
+    const context = new SsgContextImpl<MyVars>("fr")
     expect(context.name).toBe(SsgContextImpl.DEFAULT_NAME)
     const hierarchicalName = SsgContextImpl.DEFAULT_NAME
     expect(context.logger.name).toBe(hierarchicalName)