+++ src/step/content/ContentStep.ts
@@ -28,10 +28,20 @@ export type ContentStepResult = {
   contentCount: number
 }
 
+/**
+ * A SsgStep that can perform replacements in files' contents.
+ */
 export class ContentStep<C extends SsgContext = SsgContext> implements SsgStep<C, ContentStepResult> {
-
+  /**
+   * Logger name
+   */
   readonly name = "content"
 
+  /**
+   *
+   * @param contents The content roots and associated replacements to perform.
+   * @param output The function that writes the output contents once they are ready.
+   */
   constructor(protected contents: ContentStepConfig<C>[], protected output: OutputFunc) {
   }
 
@@ -46,32 +56,33 @@ export class ContentStep<C extends SsgContext = SsgContext> implements SsgStep<C
   }
 
   protected async processRoots(context: C, contentsConfig: ContentStepConfig): Promise<number> {
-    let contentCount = 0
+    let fileCount = 0
     for (const contentsRoot of contentsConfig.roots) {
-      contentCount = await this.processRoot(context, contentsRoot, contentsConfig, contentCount)
+      fileCount = await this.processRoot(context, contentsRoot, contentsConfig, fileCount)
     }
-    return contentCount
+    return fileCount
   }
 
   protected async processRoot(context: C, contentsRoot: string, contentsConfig: ContentStepConfig,
-                              contentCount: number): Promise<number> {
+                              fileCount: number): Promise<number> {
     context.debug("Processing root", contentsRoot)
     const contentFiles = await glob(contentsRoot)
     for (const filePath of contentFiles) {
-      contentCount = await this.processFile(context, filePath, contentsConfig, contentCount)
+      fileCount = await this.processFile(context, filePath, contentsConfig, fileCount)
     }
-    return contentCount
+    return fileCount
   }
 
   protected async processFile(context: C, filePath: string, contentsConfig: ContentStepConfig,
-                              contentCount: number): Promise<number> {
+                              fileCount: number): Promise<number> {
+    context.debug("Processing file", filePath)
     context.inputFile = HtmlSsgFile.read(context, filePath)
     context.outputFile = contentsConfig.getOutputFile(context)
     for (const replacement of contentsConfig.replacements) {
       context.outputFile = await replacement.execute(context)
     }
-    contentCount++
+    fileCount++
     await this.output(context, context.outputFile)
-    return contentCount
+    return fileCount
   }
 }