+++ src/step/content/replace/angular/AngularExpressionReplaceCommand.test.ts
@@ -1,0 +1,33 @@
+import {AngularExpressionReplaceCommand} from "./AngularExpressionReplaceCommand"
+import {testUtil} from "../../../../../test/TestUtil"
+
+describe("AngularExpressionReplaceCommand", () => {
+
+  test("replace text", async () => {
+    const context = testUtil.newHtmlContext("Contact.html",
+      `my name is {{name}}`)
+    context.setVar("name", "Jérôme")
+    const command = new AngularExpressionReplaceCommand()
+    await command.execute(context)
+    expect(context.outputFile.contents).toBe(`my name is Jérôme`)
+  })
+
+  test("replace numbers", async () => {
+    const context = testUtil.newHtmlContext("Contact.html",
+      `The price is {{price|number}}`)
+    context.setVar("price", 39500)
+    const command = new AngularExpressionReplaceCommand()
+    await command.execute(context)
+    expect(context.outputFile.contents).toBe(`The price is 39 500`)
+  })
+
+  test("replace context value", async () => {
+    const context = testUtil.newHtmlContext("Contact.html",
+      `The context is {{$context._name}}`)
+    context.name = "My context"
+    context.setVar("price", 39500)
+    const command = new AngularExpressionReplaceCommand()
+    await command.execute(context)
+    expect(context.outputFile.contents).toBe(`The context is My context`)
+  })
+})