+++ src/step/content/replace/html/StringEchoVarReplaceCommand.ts
@@ -1,5 +1,4 @@
 import {RegexReplaceCommand} from "../RegexReplaceCommand"
-import {HtmlVarProp} from "../../../../HtmlSsgContext"
 import {RegexReplacer} from "../RegexReplacer"
 import {VarRegexReplacer} from "./VarRegexReplacer"
 import {StringContextHandler} from "./StringContextHandler"
@@ -7,7 +6,7 @@ import {SsgContext} from "../../../../SsgContext"
 
 export class StringEchoVarReplaceCommand<V = any, C extends SsgContext = SsgContext<V>> extends RegexReplaceCommand<V, C> {
 
-  constructor(protected varName: HtmlVarProp<V>, protected defaultHandlers: StringContextHandler[] = []) {
+  constructor(protected varName: string, protected defaultHandlers: StringContextHandler[] = []) {
     super(new RegExp(`\\$\{${String(varName)}\}`, "gs"))
   }